// src/app/models/productDetail.ts
export class ProductDetail {
    id: number;
    menuId: number;
    menuName: string;
    subMenuId: number;
    mainMenuName: string;
    name: string;
    description: string;
    ingredients: string;
    imagePath: string;
    receipe: string;
    rate: number;
    discount: number;
    servePeoples: number;
    quantity: string;
    isSpecial: boolean;
    isVeg: boolean;
    isActive: boolean;
    guid: string;
    images: string[]
}